more work on CLI

+++ curator-x-rpc/src/main/java/org/apache/curator/x/rpc/CuratorEntry.java
@@ -66,7 +66,7 @@ public class CuratorEntry implements Closeable
         }
     }
 
-    public RpcCuratorEvent pollForEvent(int maxWaitMs) throws InterruptedException
+    public RpcCuratorEvent pollForEvent(long maxWaitMs) throws InterruptedException
     {
         if ( state.get() == State.OPEN )
         {
